Fixed a static body collision bug where a rigidbody could attach to the bottom of a platform and get stuck if going in a certain direction. Also added way more to the game loop for the dochs game.

+++ src/box_collider.cpp
@@ -64,7 +64,7 @@ bool BoxCollider::CollidesWith(BoxCollider *other) {
   return true;
 }
 
-const sf::Vector2f BoxCollider::topNormal = {0.0, -1.0};
-const sf::Vector2f BoxCollider::bottomNormal = {0.0, 1.0};
-const sf::Vector2f BoxCollider::leftNormal = {-1.0, 0.0};
-const sf::Vector2f BoxCollider::rightNormal = {1.0, 0.0};
+const sf::Vector2i BoxCollider::topNormal = {0, -1};
+const sf::Vector2i BoxCollider::bottomNormal = {0, 1};
+const sf::Vector2i BoxCollider::leftNormal = {-1, 0};
+const sf::Vector2i BoxCollider::rightNormal = {1, 0};
