+++ src/box_collider.hpp
@@ -29,7 +29,7 @@ class BoxCollider {
         PyObject* GetCallback();
         void SetCallback(PyObject *new_callback);
 
-        static const sf::Vector2f topNormal, bottomNormal, leftNormal, rightNormal;
+        static const sf::Vector2i topNormal, bottomNormal, leftNormal, rightNormal;
 
         bool isTrigger = true;
         bool freed = false;
